BC: Tested NuPyCEE with new integration scheme (from Andres Yague) and new import system
